Crate ethereum_light_client

Ethereum Light Client Verifier

The verifier module verifies Message objects by verifying the existence of their corresponding Ethereum log in a block in the Ethereum PoW network. More specifically, the module checks a Merkle proof to confirm the existence of a receipt, and the given log within the receipt, in a given block.

This module relies on the relayer service which submits import_header extrinsics, in order, as new blocks in the Ethereum network are authored. It stores the most recent FINALIZED_HEADERS_TO_KEEP + DescendantsUntilFinalized headers and prunes older headers. This means verification will only succeed for messages from finalized blocks no older than FINALIZED_HEADERS_TO_KEEP.

Usage

This module implements the Verifier interface. Other modules should reference this module using the Verifier type and perform verification using Verifier::verify.

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

EthereumDifficultyConfig

Describes when hard forks occurred in Etheerum Mainnet based networks that affect difficulty calculations. These values are network-specific.

EthereumHeader

An Ethereum block header.

StoredHeader

Ethereum block header as it is stored in the runtime storage.

Traits

WeightInfo

Type Definitions

Submitter