Trait ethereum_light_client::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type DescendantsUntilFinalized: Get<u8>;
    type VerifyPoW: Get<bool>;
    type WeightInfo: WeightInfo;
    type UnsignedPriority: Get<TransactionPriority>;
    type UnsignedLongevity: Get<u64>;
    type ImportSignature: Verify<Signer = Self::Submitter> + Decode + Member + Encode + TypeInfo;
    type Submitter: IdentifyAccount<AccountId = Self::AccountId> + Decode + Member + Encode + TypeInfo + Clone;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type DescendantsUntilFinalized: Get<u8>

The number of descendants, in the highest difficulty chain, a block needs to have in order to be considered final.

type VerifyPoW: Get<bool>

Determines whether Ethash PoW is verified for headers NOTE: Should only be false for dev

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet

type UnsignedPriority: Get<TransactionPriority>

A configuration for base priority of unsigned transactions.

type UnsignedLongevity: Get<u64>

A configuration for longevity of unsigned transactions.

type ImportSignature: Verify<Signer = Self::Submitter> + Decode + Member + Encode + TypeInfo

type Submitter: IdentifyAccount<AccountId = Self::AccountId> + Decode + Member + Encode + TypeInfo + Clone